Use the new hash-detector-v2 for the request reporter.
