--- reporting/src/request/hash/prob.js
@@ -1,601 +1,0 @@
-/**
- * WhoTracks.Me
- * https://whotracks.me/
- *
- * Copyright 2017-present Ghostery GmbH. All rights reserved.
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0
- */
-
-export default {
-  'logM': [
-    [
-      -5.317193109695505, -3.913874405861832, -2.9381665079978383,
-      -3.191773525248805, -3.7998673319383616, -4.345439435400599,
-      -3.2318181091222926, -4.0115754219957696, -3.6177876283722235,
-      -7.304959588802769, -4.086587827961371, -2.4934668495341517,
-      -2.900130610423348, -2.0358548114690445, -6.470340186675337,
-      -3.8789458371913224, -6.60892942891667, -2.30640750219933,
-      -2.92314951300914, -2.2528985979367016, -2.813060976891666,
-      -5.056683799455272, -5.62960615098086, -5.242592116855742,
-      -3.535964746434304, -3.9784518141894374, -6.985122896496697,
-      -7.274967682087109, -7.9594919332349345, -6.164538915616592,
-      -8.320505278772265, -8.280727307315345, -8.550921589415555,
-      -8.54909843485404, -7.934552984887683, -8.127372805874096,
-      -3.109412750730325, -3.90126314517224, -11.219610716637185,
-    ],
-    [
-      -1.817961182029586, -4.477008918746111, -3.982294685743666,
-      -5.596742628697303, -1.3277338286038127, -5.81319091362829,
-      -5.706241389831284, -4.704855751509904, -2.3707799708377664,
-      -6.688098536219502, -5.946054348175743, -3.0303647822952886,
-      -4.927999302824114, -5.459013629433194, -2.0897428400753895,
-      -6.037348386957373, -8.202702345982763, -3.0256690344728256,
-      -3.7332038394562588, -5.308746692899046, -2.4949707632290035,
-      -6.088580077804754, -5.296210276492634, -8.130539899734533,
-      -5.057641495903164, -6.320488830892931, -6.656443335170194,
-      -6.082438809782672, -6.9118032022964, -7.14931515531959,
-      -7.6756820362967915, -5.536088462601377, -7.86768889882521,
-      -7.907396348420323, -7.959356087351034, -7.501087714323719,
-      -3.2616674937184142, -4.221686469067666, -11.94037196426613,
-    ],
-    [
-      -3.1504260003263047, -5.8946485140234195, -4.462130109094279,
-      -5.654975660757999, -2.6862615330360584, -6.254615079727427,
-      -6.115732555639961, -1.0345299818896994, -4.339707120241444,
-      -8.364050977275454, -2.9637094236602577, -4.067987318306986,
-      -5.314777936793433, -5.937887707129599, -1.1763414859058363,
-      -6.325832151419573, -8.626415241742945, -4.2703048846988825,
-      -4.649478785472155, -3.572869468862069, -4.49847473009233,
-      -7.7949564453854885, -7.003074424139854, -8.55875659326913,
-      -5.4163783655419495, -6.99586836609861, -6.300439195958324,
-      -6.140398097670918, -4.296032840221402, -7.198499205932236,
-      -7.707461392449307, -8.361490156413781, -7.774840824588481,
-      -7.207331057800064, -8.254851685310463, -7.607567921543699,
-      -4.452335711501991, -4.652168893134668, -12.384287567343833,
-    ],
-    [
-      -3.162725777484131, -5.217203426524381, -5.551773316344876,
-      -5.225969121348031, -0.49466356298299, -4.277110633369936,
-      -5.803309498701423, -5.115021553085183, -2.3928970315355627,
-      -7.271882693744789, -5.616223149619324, -5.101503026330162,
-      -5.2502736229337845, -6.030238363252401, -3.3753725823208587,
-      -4.7569364291327325, -8.75633959600018, -3.910798063850956,
-      -4.10691723970344, -4.33316315169083, -3.5462564103937786,
-      -5.795781588710692, -5.60645664261893, -8.160690779925112,
-      -5.5179826355601564, -7.002162837921167, -5.66318809155304,
-      -7.047625211997925, -7.459294583710373, -7.8790510219292935,
-      -7.686366678099793, -7.89740016059749, -7.970618215454785,
-      -8.252092943332231, -7.545830820694459, -8.350874487892016,
-      -3.7994740567963308, -3.2784138216874594, -12.09402357194001,
-    ],
-    [
-      -3.7189886827898455, -3.136327618286749, -3.758404308148985,
-      -3.1734627642603788, -4.270043208702629, -3.9851745789702915,
-      -4.073191269596321, -4.243313073155838, -2.7366356984012534,
-      -6.69301445301315, -4.289361094233771, -2.9225173480317617,
-      -3.7994772042488547, -1.9820771807446453, -4.744874009923719,
-      -4.076925681216796, -7.571527682479396, -1.8537894256636591,
-      -2.6732861064668727, -3.07540677713879, -3.8623600914903835,
-      -4.956475298209049, -4.087386942308493, -4.097862478306168,
-      -6.172798016825123, -5.615949932671272, -6.334419930219192,
-      -6.653374688453774, -7.6080871143035695, -7.667839399208454,
-      -8.064223854680872, -8.372070201736241, -8.654386220683383,
-      -8.289488297564219, -8.649302325646111, -7.949080835643628,
-      -2.4290509512920795, -2.9832968468317835, -11.736296190478427,
-    ],
-    [
-      -1.8610170544945401, -5.020582941750675, -4.821798725236716,
-      -5.14763132949712, -1.9970397814929628, -2.9637268592690686,
-      -5.4773413442463905, -5.226847392244456, -2.1516913692262554,
-      -7.139830733830931, -5.003454059487709, -3.071744448015679,
-      -4.70568319038359, -4.537956069904344, -1.961246272242613,
-      -5.7210740965663796, -8.830298869007885, -2.386838772856728,
-      -4.40737684801158, -2.9024924967110852, -2.8901517595050215,
-      -6.133318891659922, -5.897516165575739, -7.453221162992919,
-      -6.484619136550752, -5.488961875269409, -6.440778854546247,
-      -6.356926372465171, -6.85596303345462, -6.983506927469359,
-      -7.1954972780819295, -6.954974321433788, -7.412399168472664,
-      -6.290359746909414, -7.331331345383882, -7.597390953822497,
-      -3.640302563386591, -3.4894582036851673, -10.749058028997247,
-    ],
-    [
-      -2.670791162329886, -4.801794512175134, -5.370352088589672,
-      -4.771716320175903, -1.3789480193077568, -4.985947863046129,
-      -4.867460915336402, -4.1641163040595135, -2.9077027799919124,
-      -7.340436895180519, -5.469216169986365, -2.2755965131414464,
-      -3.7442985057391613, -4.534865389130813, -2.166823499104593,
-      -4.579327537409528, -8.437790530441433, -3.186109987787664,
-      -3.298269839791393, -4.600572401777362, -3.3763664717270916,
-      -5.467859974873916, -2.96249605191553, -7.974887522745274,
-      -5.6773030354806195, -6.235130020840195, -6.231799459893031,
-      -6.453800688013288, -7.568630962280797, -7.862662262411237,
-      -7.8928369655334505, -8.557934842283496, -8.426533375916797,
-      -8.775773341368012, -8.508132019799678, -7.68958357269846,
-      -2.792677043930547, -2.9092902958635007, -11.309470155325444,
-    ],
-    [
-      -2.2482601312819104, -5.126347808963377, -5.7772675133943325,
-      -4.69443449591098, -1.6533205151057395, -5.850684281096827,
-      -5.961298395421176, -5.748036025541178, -3.0296509371927796,
-      -6.940148696631176, -5.3943774114013525, -3.4425959982219547,
-      -4.336900017361096, -3.2755737205206654, -2.2900711271934324,
-      -2.9452842712742293, -7.154167933816175, -2.9876200659170675,
-      -4.121847012655317, -1.604793914367351, -3.462265449036904,
-      -6.3092899943704515, -4.502909548838425, -8.340196441388679,
-      -5.352513973469052, -5.84513742013666, -7.716458712200901,
-      -7.503365496740193, -8.214347544936173, -8.636663343396192,
-      -9.468396687242285, -9.02463529405684, -9.208113588978618,
-      -9.065939100198076, -9.14097428614099, -8.193990607867429,
-      -3.446652002735739, -3.471160539621863, -12.568488976120518,
-    ],
-    [
-      -3.0007497362890985, -4.443273506683785, -2.600406125797639,
-      -3.5506673813608214, -2.3088581699376767, -4.3884112328791565,
-      -3.3240861733092757, -5.410693842604261, -6.299225959971416,
-      -6.948909412333564, -2.6175027037374363, -2.6781627546892772,
-      -3.4886131234798126, -1.599774904001608, -3.3606225411317303,
-      -3.247255852518435, -7.115193970217702, -3.7280955129350484,
-      -2.6923907966021092, -2.555377813224534, -5.807844434815872,
-      -3.9345110232700735, -6.202911259741086, -5.561268767713861,
-      -6.650031586366622, -4.712928097133364, -8.304128897179098,
-      -8.244734179760869, -8.889828189920294, -9.755160804215802,
-      -9.131234000547042, -9.921057481342356, -10.257529717963568,
-      -9.736232794330283, -9.278017120836589, -9.921057481342356,
-      -4.768461508802108, -4.254995091958906, -12.17234927994885,
-    ],
-    [
-      -1.4940919539963653, -5.922446524399309, -5.289777262802165,
-      -5.624256474349113, -1.8851775294076067, -5.715394186347637,
-      -6.021048501203849, -5.908460282424569, -3.396905691153677,
-      -6.504615196997205, -5.1924288080310355, -6.3420962674994295,
-      -5.674266894923774, -5.49377851921148, -1.2976578392687523,
-      -4.257311881409098, -5.917762675086882, -5.734662605213514,
-      -2.8117520915305754, -5.762277772246487, -1.8476980380366512,
-      -5.307387780810799, -6.036714617948248, -7.7872311286422535,
-      -6.783929019778468, -6.349264756978042, -6.096352884267978,
-      -6.889289535436295, -7.223295679562314, -6.208564874874907,
-      -7.391918391998107, -7.172864825935423, -8.450525346052517,
-      -7.570166623404426, -7.700219751652623, -6.7404439078387295,
-      -4.628298339741086, -4.624444770425097, -9.674300777674633,
-    ],
-    [
-      -2.250334306208922, -4.933230325278283, -5.466394927126537,
-      -5.398473579985344, -2.2268113016383833, -4.748334764489155,
-      -5.788809806497209, -5.0078142235138055, -1.6140601096159817,
-      -7.389966098099587, -5.057421922662515, -2.898611313783839,
-      -5.522933405597031, -5.010356328473512, -2.4628493813340273,
-      -5.379885684217253, -9.141316028989765, -3.3594119596664482,
-      -3.041463375786868, -2.419747191598849, -2.9442671356242585,
-      -5.592780369662592, -5.212796184399939, -8.509412473670539,
-      -4.991914636312223, -6.204724330479714, -7.0945818052857135,
-      -6.033759029363022, -8.23098360669924, -7.725751285569761,
-      -8.376539192290565, -8.49374635692614, -9.161316695696435,
-      -8.63844273341427, -9.141316028989765, -5.336635700423437,
-      -2.320068203582952, -3.2058981086279252, -10.760704272277033,
-    ],
-    [
-      -2.1974731235858997, -4.3388628455791665, -4.967718897078009,
-      -3.2621688771724497, -1.4308113982798378, -4.213389336390515,
-      -4.802169512504785, -5.329295377389805, -1.9331585330356604,
-      -7.11868966103088, -4.521078867045072, -2.582112507472284,
-      -4.200019071522844, -4.723762515330896, -2.5458289029974734,
-      -4.394052068089149, -8.833409345865157, -5.542835649905102,
-      -3.525590022142182, -3.219261585927306, -3.440694886256806,
-      -5.260002606130265, -5.707719973255917, -8.134374022679989,
-      -4.815821121322918, -5.001435609811865, -6.751510196656888,
-      -6.440177758649728, -7.115152338630028, -8.064468325935678,
-      -7.64931791209305, -8.538861135600177, -8.306958404196084,
-      -9.077438788159201, -8.079174473325374, -8.31395959965507,
-      -3.143648602335109, -3.23841992629694, -11.375011339329703,
-    ],
-    [
-      -1.870891141376203, -3.7591745118646447, -4.371273060301154,
-      -3.2791332861386615, -1.9107706682716499, -4.624061777758882,
-      -4.956148042436841, -5.010189937826345, -2.784451177890736,
-      -6.334089349722556, -5.299830748963299, -2.0044177702017865,
-      -3.1473482277763787, -5.127479879790951, -2.673071856202927,
-      -3.240994097679745, -6.869939166060555, -4.868141840731485,
-      -3.233492162505462, -4.180898334655049, -3.485611756924462,
-      -4.7287102962346195, -3.225464419775572, -4.446417109530032,
-      -4.675740725834205, -6.102661940484367, -6.8918082037979795,
-      -6.5504745964216236, -7.780375621470422, -7.458715195376792,
-      -8.381149481899353, -8.51588207586951, -8.729456176167568,
-      -8.721550996660454, -8.733432324547207, -8.690540759917894,
-      -3.3429426189259286, -3.357692904656495, -11.956300170684946,
-    ],
-    [
-      -2.931405039512769, -3.8790732080718975, -4.208598923580739,
-      -2.1460310544551073, -2.0679892632481707, -3.8710594894771995,
-      -2.317196987858577, -4.558600992200827, -3.129123860689175,
-      -6.201820607270832, -3.4226514708493125, -3.4916353014100454,
-      -4.8054768369101755, -3.8517699588791987, -3.587068773614783,
-      -4.65485761895647, -8.027890699333087, -4.962007805103018,
-      -2.9316585302933995, -2.740903095629866, -4.065599090041162,
-      -5.257029590537684, -4.870572036830945, -6.895709352566281,
-      -5.060638408544645, -3.818366324945421, -6.598521417385409,
-      -6.123432914955803, -7.4421519040830475, -8.0244444006895,
-      -8.313656695322049, -7.836785431300824, -7.908702182067969,
-      -9.014490308806149, -8.476265710852045, -8.108125216731965,
-      -2.395207538044643, -2.5176780774464413, -11.537137286576998,
-    ],
-    [
-      -4.582048172975939, -3.6721608040964133, -4.013673492134655,
-      -3.6768208008095375, -3.683218218775885, -4.040490497653242,
-      -2.8457319454164787, -4.659595811215548, -5.290561241831474,
-      -6.533679855748992, -3.3658001316957837, -3.333626612775081,
-      -1.7116898687428947, -2.2145540058139983, -2.5876318074065163,
-      -3.5857620783742234, -9.233970561080291, -2.041442328336181,
-      -3.235672531430412, -3.3788363510457717, -3.013819880171509,
-      -4.581149576108947, -3.8712335663317337, -5.071922202638369,
-      -5.807727147492048, -5.946082541240341, -7.605156369683052,
-      -6.784716119911276, -8.590973164205817, -8.865410009907578,
-      -9.817418824383811, -9.451694044925162, -8.627834757509975,
-      -10.034089861192404, -9.964022298575687, -9.39053962177183,
-      -3.6613757818291144, -4.031896780839415, -12.215314097182182,
-    ],
-    [
-      -2.2729171158438817, -5.285919434448688, -4.6696834973756935,
-      -3.4923420438473705, -2.005653594665365, -3.9271504148616305,
-      -5.967450755562108, -2.467301079019551, -2.669148729684826,
-      -7.333133556946934, -5.6270375739531495, -2.8463660347468736,
-      -5.229687197958054, -6.062723071637988, -2.0304045709327947,
-      -2.9545066115874072, -8.904721245249936, -2.0999087720223852,
-      -3.1552145260461386, -3.8745167778097698, -3.6803195616520683,
-      -5.739547628912189, -6.22467866913463, -4.787195083684797,
-      -5.8804709327747124, -6.0757184337300645, -6.564290009765776,
-      -6.376709494927113, -6.038447038932833, -7.098777141029926,
-      -7.479893220127803, -7.87255405598998, -8.328910771394533,
-      -8.222346161671902, -8.681577693935727, -8.081172565074397,
-      -3.4561997517901255, -4.009603697373822, -12.23692575542514,
-    ],
-    [
-      -3.873157429206279, -4.830270155600689, -4.43527634735982,
-      -4.14496615250177, -5.111031785207897, -5.025170494601504,
-      -5.5551660344752145, -4.63672707406277, -3.2260541871477404,
-      -5.770277414092161, -5.681046280364217, -3.716153680396614,
-      -4.740657996911002, -5.610428713150264, -4.55437603890095,
-      -4.152871332008883, -5.227953123266799, -4.9344739331306915,
-      -4.11395591575921, -3.788816280772528, -0.5653919923366761,
-      -3.6759472599185736, -4.500516869228221, -5.56597695057943,
-      -6.532417466139057, -5.656948728785157, -4.969771715211715,
-      -4.493054148026632, -5.227953123266799, -5.770277414092161,
-      -5.5027980489578985, -5.853659023031211, -5.599129157896331,
-      -5.621857408973887, -6.686568145966315, -6.303575893710209,
-      -2.8328807175511326, -3.2115009157377044, -8.47832761519437,
-    ],
-    [
-      -2.3501201454598113, -3.877286819275504, -4.035451330252564,
-      -3.7231361394482456, -1.995970055224952, -4.265044156081665,
-      -2.736694869216493, -4.688450598496515, -2.5992406567264554,
-      -7.320698648200759, -3.583928454056477, -3.8563530232043615,
-      -3.7853331501533014, -3.214809130719206, -2.5959281587718417,
-      -4.532099917781923, -8.54022004727565, -4.417618701124041,
-      -3.048279251990644, -2.5942436477724358, -3.1242479465432065,
-      -4.469309439914175, -4.806633372411803, -7.394619249171621,
-      -4.555828431285585, -4.358563465976057, -6.866501378881329,
-      -6.331832351396451, -7.673379491145088, -8.197874078642343,
-      -8.347641990314505, -8.913227945236432, -8.569510844196715,
-      -8.740432657525941, -8.47054012663766, -8.548501668107372,
-      -2.9529559443829054, -2.699051077505382, -10.812345932784986,
-    ],
-    [
-      -3.232187843032739, -4.0820117364975985, -2.303935084897713,
-      -4.462431762002812, -2.090040838747268, -4.810263669875053,
-      -5.098149818280902, -3.3427415006907157, -3.0890624986145476,
-      -7.01228575724969, -4.1419364125858, -4.24045222699042,
-      -4.229931087854754, -5.100601983730359, -3.4039828305010356,
-      -2.8099921854072107, -7.214019479079884, -4.907194618144589,
-      -2.643507033442835, -1.7540357631259802, -3.3854048552772955,
-      -5.184041819845991, -4.47680113344124, -7.6899519794478985,
-      -4.808040087881555, -5.219185131072912, -6.5678322420137585,
-      -6.167867875813282, -6.8402501017818516, -6.916380979124739,
-      -7.5678834584129655, -7.658729931186539, -7.2613021593846385,
-      -8.191570040629912, -8.226544462901742, -6.8015999555430176,
-      -2.6010778450351553, -2.6904387628877977, -11.662228020635066,
-    ],
-    [
-      -2.5098219012317164, -4.866751265554057, -3.6932794196268826,
-      -4.771537513115893, -1.6104104195979998, -4.786730809524362,
-      -4.844014888999637, -3.3324721430935713, -2.6555999682102924,
-      -6.936119330196025, -4.965055025885696, -4.38862097481251,
-      -2.432270482023462, -5.312502160603078, -2.847025899013536,
-      -4.704861692289055, -9.182649668516397, -3.038045783139583,
-      -2.9173621825204403, -3.3068877414035236, -2.6889941420441494,
-      -4.419119006921773, -4.457628790520158, -6.774956799381695,
-      -4.746475063409969, -4.034966308526627, -6.2439271683315285,
-      -5.181571647712297, -7.664630599996934, -7.674169623043693,
-      -7.670979827675593, -8.475165324810046, -8.85602462211601,
-      -8.257016432469927, -9.032115629159328, -8.277248336441511,
-      -2.8651461842364894, -2.981401171615819, -12.609442597367517,
-    ],
-    [
-      -4.82838528988538, -2.5104681625450804, -2.951707407664075,
-      -3.9834528948235115, -2.3794671622228702, -3.722207396872334,
-      -4.073477470440828, -4.736386791191092, -4.61444776796882,
-      -7.109351077520736, -4.004146193053924, -3.3735707029212727,
-      -2.8270580994946104, -1.6595892968356591, -6.219331412258552,
-      -3.6728896449692385, -8.956279393027199, -2.4527988314729923,
-      -2.3891799588738047, -1.9700952733996555, -6.565069913996628,
-      -6.177167490448294, -6.0378013973162945, -6.0787883752796565,
-      -6.7157925201514495, -5.626883610473226, -8.156463652149412,
-      -7.909603574217885, -8.514820682271129, -9.603382635085737,
-      -9.4294293279623, -9.364365734615237, -9.675953327920572,
-      -9.891064707537518, -10.16550155323928, -9.231819078653254,
-      -4.345314510445386, -4.3673662947350795, -11.905967728079784,
-    ],
-    [
-      -2.551610846439679, -4.59048814225872, -5.259589857964013,
-      -4.259853450863533, -1.1156080256590044, -5.401623465644263,
-      -4.935087379845991, -5.400393453189063, -1.3071455789769282,
-      -7.1854147652231415, -5.445012855669765, -5.087823905037628,
-      -5.008750122625029, -5.1547633877127375, -2.169032888642808,
-      -4.896846869643483, -8.317549745929895, -4.666006418282462,
-      -4.2373194748535905, -5.115635134455043, -5.417752847574147,
-      -5.408415711578208, -5.246330581238229, -6.892253227006735,
-      -6.129202842625693, -5.637151076158195, -6.570328131132742,
-      -6.203212828399443, -7.0234454372780855, -7.215156734421879,
-      -7.732291527381134, -7.42891054538625, -7.7514614434888545,
-      -7.659088123357839, -8.28402705389125, -7.585400407566681,
-      -3.1725689967085313, -3.4451321326492406, -11.185448647974,
-    ],
-    [
-      -3.5382996343990234, -7.1760564396779465, -7.488247397543631,
-      -6.978396994387756, -3.398439769465693, -8.09689706565408,
-      -7.810641091029813, -6.264185317233577, -2.968073461539628,
-      -9.363600399303442, -7.725108891071505, -7.288866452774152,
-      -7.557968428244685, -5.535336053029335, -4.5411070357698415,
-      -6.77470700942525, -11.082904789316075, -7.3709075260868335,
-      -4.656599039852368, -5.964331658547692, -6.005918102198901,
-      -7.595727027909167, -0.579558571288889, -10.350226769999495,
-      -8.488613473570524, -8.30528840439933, -7.625774355972108,
-      -7.973699831606823, -8.575405816499481, -9.312885306724427,
-      -9.398898181384466, -10.261279283982999, -10.297384288625114,
-      -9.708780963816439, -11.184687483626018, -10.039555179323015,
-      -1.2639829265502496, -5.643283892252112, -13.600601261927066,
-    ],
-    [
-      -3.951201982257603, -4.948319428117503, -3.8408143360232323,
-      -4.127704956118702, -3.401870040429183, -4.1851341708539485,
-      -4.22681983296669, -3.693877198102415, -2.386139109555554,
-      -7.1165564685804705, -5.351825671740334, -4.565327175861397,
-      -4.402429530206852, -4.592429105639189, -4.457696511669032,
-      -2.7633703227684006, -6.892270712687483, -5.932203451124186,
-      -4.564436307205596, -2.7602140093615457, -4.842958912459677,
-      -4.597938761450158, -5.194604507993407, -2.1480138172059253,
-      -3.910100306572051, -6.664571344837413, -5.3372552342744575,
-      -5.347919416773269, -5.5696020474478045, -6.346448246884397,
-      -6.8696963906489446, -6.594367086164165, -7.044900479674035,
-      -6.747952953776464, -7.35408875734678, -6.883179740986232,
-      -1.1390604573665573, -2.768456016427479, -9.642285112888725,
-    ],
-    [
-      -3.0360878131472138, -3.975479876294767, -3.522670425065406,
-      -3.914638217038522, -2.864136346767505, -4.612425777154969,
-      -4.692904626436059, -4.686936334807632, -4.041084976685928,
-      -6.232090221372901, -4.721994099636263, -3.71645641175006,
-      -3.413582384868336, -3.715626093215908, -1.439131316676773,
-      -3.4289523544300353, -7.025552744548839, -4.153383163751057,
-      -2.8472271859566614, -3.8317949474225292, -4.996465089014618,
-      -5.35198223751004, -5.0316478708463475, -7.279188975590903,
-      -6.915010870149015, -5.547052189562215, -6.037847832498722,
-      -6.144956482804655, -7.540716769913428, -7.951631482789157,
-      -7.765219940743677, -8.191304336054577, -8.140532010681154,
-      -7.842997641786361, -8.994452411828851, -8.019454079127918,
-      -1.8661802841111401, -1.988283949929875, -10.047602326420204,
-    ],
-    [
-      -2.927583140905028, -3.7903328441395248, -5.2764489864630795,
-      -4.2312936047344305, -1.1559880630047743, -4.676272330389096,
-      -5.23287773816944, -4.767939518914919, -2.360571468321946,
-      -6.515141371061506, -5.032358065889634, -4.165573658508511,
-      -5.198053762297738, -5.042834135281574, -2.0300547691548245,
-      -5.266664735746823, -7.98681534378312, -5.87003613166564,
-      -4.661779323086529, -3.113677890408956, -2.513120766550449,
-      -5.334185936622318, -3.8859702169579258, -8.799921839422044,
-      -5.109572877328748, -4.437697832044343, -6.624724584404116,
-      -7.034653985640966, -7.280709018580669, -7.53879062125616,
-      -8.00435965343403, -6.735295383527349, -6.9696960893661934,
-      -8.20995889509733, -7.995549023751875, -7.566961498222856,
-      -2.8709586741433575, -2.9027679717853037, -10.785837323091057,
-    ],
-    [
-      -4.5775657751997905, -4.58992365604639, -5.367705221747046,
-      -5.055747550889418, -4.967954414902319, -5.149703069933254,
-      -5.805746500289823, -6.008135620134587, -5.960633286149583,
-      -7.194250807060516, -5.562064757100479, -5.347958458812211,
-      -5.83941971539641, -6.072313845825997, -6.5793043738080454,
-      -5.7499380283619415, -7.970394907191429, -6.262156410013235,
-      -4.791540115647279, -4.74691431593648, -3.4961222229285234,
-      -6.555662610751005, -6.416307089840494, -6.1947908781856595,
-      -8.137448991854596, -6.679891610713214, -2.7802196992630956,
-      -2.466158404610568, -3.0417486580993924, -3.02177141454927,
-      -2.7040247698301756, -2.4146812172873138, -3.502499374832007,
-      -3.2062505284063745, -3.0890509233383026, -2.0082007423835604,
-      -1.9175368032041353, -2.586870675421622, -10.312200713338756,
-    ],
-    [
-      -4.921223474189089, -4.633780664808195, -5.167356543727998,
-      -4.784586859912606, -5.003981187922294, -4.974281104258012,
-      -5.405296912483647, -6.479368092235639, -5.3348558148203304,
-      -7.762121774414689, -5.227904669082356, -4.539319242930317,
-      -5.096488646215818, -6.355542742123293, -4.81682035448759,
-      -5.405296912483647, -7.6833408965615755, -5.624952764079572,
-      -5.278582178838222, -5.650419370516632, -6.5597896595462135,
-      -5.779425868542865, -5.973098285005509, -4.457820528974706,
-      -7.847643947852851, -7.516286811898409, -3.139562672174348,
-      -2.809879986368343, -3.3729486060067337, -1.5438156042637392,
-      -3.314852460842809, -3.441578224193167, -3.6923946909935523,
-      -3.36202174698568, -3.7087699863221735, -1.5359739580161158,
-      -2.537345075646371, -2.5622670925805493, -10.2635577261539,
-    ],
-    [
-      -2.6379446880572806, -1.8895218268654912, -5.130742122338827,
-      -3.6507176890243387, -4.63168566715247, -4.895604969143138,
-      -5.535293571178316, -5.663749801630792, -5.119832923238473,
-      -7.117702811224498, -5.818419827094236, -5.094831621033055,
-      -5.034933479451987, -6.0645528966331455, -6.305292209837734,
-      -4.891279079196016, -6.935381254430543, -5.411743331799775,
-      -4.677096420333077, -5.154795085800613, -6.59890901780933,
-      -5.471611581142362, -5.929859388828445, -6.559993601559657,
-      -7.483947206179381, -7.131125831556638, -3.0812227944475366,
-      -3.110854592053908, -3.1324293440567588, -3.3680418848507023,
-      -3.1575247639093043, -2.993368040369347, -3.438200518923158,
-      -3.3137105500100565, -2.897999046837881, -2.6189307160262256,
-      -2.2814209042730194, -2.350300403502041, -9.489280775705495,
-    ],
-    [
-      -4.719359742571733, -4.431677670119952, -4.66010070115941,
-      -4.320018352314098, -4.691663812182475, -4.175389874571988,
-      -4.862178856389441, -4.88353198086001, -5.75412055925331,
-      -7.136984385276411, -4.58582834994721, -5.1930331007064305,
-      -4.641856181416872, -4.366967460425001, -6.623422781691546,
-      -4.757945736566923, -6.973096530171294, -4.57308932416978,
-      -4.3618304628270845, -3.8255418665496363, -4.397470914718416,
-      -5.75412055925331, -5.836456845913945, -6.085109574459991,
-      -5.527546472842311, -7.178948584375443, -3.297974061640572,
-      -3.3416491251432334, -3.3921649152801354, -3.472369553162106,
-      -3.2787090174876337, -3.53238032833323, -3.4897120511646937,
-      -3.6058007001848207, -3.111632694541262, -2.771010567917061,
-      -1.1162270641274428, -2.332177662294085, -9.818005913990701,
-    ],
-    [
-      -4.392055704741102, -4.537713277163465, -4.263724531843686,
-      -4.233764232247812, -4.531333670199426, -4.66554664867335,
-      -4.994050715645386, -5.470133390967503, -5.446035839388443,
-      -6.606017074432845, -5.365993131714907, -4.772929409711944,
-      -4.671016124477886, -5.9715186209544076, -6.163280571527449,
-      -4.748928257612401, -7.214264811206304, -5.667197229498662,
-      -4.7197148851465, -5.4580120304351585, -6.179674381303125,
-      -5.568293155439748, -5.604826898772203, -5.34784381320923,
-      -7.168802437129547, -6.4428654337466105, -2.875303479963259,
-      -2.9835344070473893, -2.9463578722801302, -3.0961734418175957,
-      -2.957827766991389, -3.104917082392155, -3.232195226070862,
-      -3.1765316168586497, -3.1926576640601723, -2.0775092400158366,
-      -2.0518737293268643, -1.8088674778716027, -8.14225158284365,
-    ],
-    [
-      -4.160429085723273, -4.311682262617893, -4.346457176794503,
-      -4.105388739787635, -5.004829443177838, -4.23027728948118,
-      -5.1606693125959415, -4.8267362224934045, -6.063734282250835,
-      -6.994731161514686, -5.4039853820927375, -3.6785269926859225,
-      -4.926931194850336, -4.600208573194954, -5.641380956114149,
-      -4.613814225250733, -7.2701431413746525, -3.717586428522509,
-      -4.060086579940533, -4.405853850357913, -6.794060466052534,
-      -5.705600548848443, -5.458262268947176, -5.5499932045247204,
-      -7.294835753965024, -5.973079913982704, -2.939088928311623,
-      -2.9536311138113973, -3.0495599745426376, -2.926654526113195,
-      -2.8183515756049773, -3.110879033053236, -3.2632535134792384,
-      -3.056390847769166, -3.2365502415586254, -2.485093402248158,
-      -2.252217414015717, -1.9042811059505762, -8.49880855829096,
-    ],
-    [
-      -4.6360984222176, -4.799462065266941, -4.404379306200994,
-      -4.322440863362559, -4.413912266859718, -4.779935642916586,
-      -5.9094059199446, -5.4193064266957265, -5.6089848217153895,
-      -6.759556849314211, -6.0968688762389736, -5.348569875603949,
-      -5.047562348555018, -6.628528586907806, -6.264860607478103,
-      -4.897028309197948, -7.375742988738027, -6.217232558488849,
-      -4.9545521533361345, -5.040368072920991, -6.842938458253261,
-      -6.086612376071785, -5.388376126004368, -6.027188955600984,
-      -6.381120413593965, -6.512696771382685, -2.7731687619307577,
-      -2.6782975266407654, -2.830480089320237, -2.848734564462583,
-      -2.7727974977720473, -2.708815924574242, -2.758058837458462,
-      -2.900512768676951, -2.978098213907181, -2.5702051835137754,
-      -2.200430601827526, -1.9914178350479794, -9.062141942308257,
-    ],
-    [
-      -4.582473439004593, -4.754993084804229, -4.998702913699916,
-      -4.321922710604176, -4.998702913699916, -4.818006052632963,
-      -5.75898939709749, -5.610808884256736, -5.659898494453259,
-      -6.539147954647065, -5.381695165956022, -5.29945706771905,
-      -4.9890097844942565, -6.113682179832232, -6.240166326941351,
-      -5.286384986151697, -6.539147954647065, -5.11567459028542,
-      -5.183625252193928, -5.041827340333671, -6.113682179832232,
-      -5.966628761875734, -5.949679203561962, -5.8767724327538735,
-      -6.63445813445139, -6.398069356387159, -2.9455786803374537,
-      -2.890853780648207, -2.785374928797118, -2.8810401591998827,
-      -2.844226186077166, -2.950591222160998, -2.8659206620320004,
-      -2.962385798653835, -2.9290493783862432, -2.175084730574881,
-      -2.0525565754026527, -1.9973049026102476, -9.11936478423939,
-    ],
-    [
-      -4.543401159590459, -4.823115062393064, -3.9860737021730492,
-      -4.4553902822677465, -4.626782768529511, -3.1798637855931853,
-      -4.774072977325461, -6.066661375783507, -6.3635601033402125,
-      -6.315358001522335, -5.921727351061174, -5.495319069107766,
-      -5.152207191716654, -5.604717663514872, -6.866188959902025,
-      -4.689583669768541, -6.951346768242331, -5.646021469656108,
-      -4.508034015753168, -5.670412922780267, -6.3635601033402125,
-      -5.774402636304314, -4.9408980980494475, -6.225409764859395,
-      -7.069129803898715, -6.845986252584505, -2.730859151179876,
-      -2.4339154965622467, -2.826006105651267, -3.0005304914284725,
-      -2.771844397679924, -2.829963657137874, -2.9877860538538656,
-      -2.8145814901362787, -2.6436845488573124, -2.654817505726865,
-      -2.429558191193291, -2.2946388732939207, -9.148571345578551,
-    ],
-    [
-      -4.725109169299543, -4.8920815045246915, -3.9074073677574606,
-      -4.618580271717583, -4.733235624856568, -5.307693562176592,
-      -4.572758669976433, -5.510513640518887, -5.762673734812137,
-      -6.741812370345956, -4.947803181383885, -4.846383002937982,
-      -4.158324891885508, -5.898382986736673, -6.406305850844702,
-      -4.702876150359534, -7.123904127268811, -5.54258051631555,
-      -4.607326133643159, -5.0810135982393225, -6.025291838600702,
-      -5.939931989649545, -5.3109193714254745, -5.302874275740643,
-      -8.128106731465847, -7.104295655880435, -1.614636052812513,
-      -3.0395098958998785, -3.254561277316752, -3.266619635483967,
-      -3.589134099707649, -3.5175456968428787, -3.536816207673623,
-      -3.603970735498501, -3.4677319911306594, -1.8912612403463043,
-      -2.455712659678065, -2.0020640424692564, -9.54180006677385,
-    ],
-    [
-      -3.0571113135980825, -3.6881540513013364, -1.8979707427255759,
-      -1.2943462356308515, -3.718142414898046, -3.529032437735793,
-      -3.056990559456546, -2.595180349879562, -4.19957318353627,
-      -5.163760465623824, -4.283245007257953, -4.389915665629348,
-      -3.7239696638225332, -3.6638309678336567, -3.339808005324408,
-      -3.2232840549585458, -7.201600756421915, -4.424436511454205,
-      -3.4318215195509154, -3.7600626508030572, -4.889716612123718,
-      -4.907633282095024, -3.3897747516398495, -5.280394676088581,
-      -3.746319271008434, -5.383591676103269, -6.004427278244468,
-      -7.145088546158573, -7.502467467599099, -7.389874232982313,
-      -8.34616042141476, -8.160485969610496, -8.582297342028976,
-      -9.363761666902276, -8.540136531204698, -7.242105372816881,
-      -9.197243775281446, -8.16678189806731, -12.826019305325678,
-    ],
-    [
-      -2.7788146555274436, -2.8216148673471526, -3.490269174255095,
-      -2.629595716118393, -3.303540064657417, -3.0594591567393,
-      -3.267058338000351, -3.3605744282568595, -3.338572680204692,
-      -4.862200431472076, -3.079621451687009, -3.4779532672383615,
-      -2.9171539676677303, -3.715133767598423, -3.489484464837551,
-      -3.1815170461700744, -6.351573294671603, -3.497698294421097,
-      -2.435128706315454, -3.021635590790534, -3.45696216959138,
-      -3.6719853295036016, -3.335125700110903, -6.06609954111505,
-      -5.887715966274562, -4.246403490039957, -4.1187166377434625,
-      -4.176821573187442, -5.030253279494612, -5.2339703619139595,
-      -5.394178255190106, -5.739040880804595, -5.797763310103027,
-      -5.845070727353406, -6.104151691418398, -5.559409206065855,
-      -8.002718669789934, -5.329333984788067, -12.431833340562715,
-    ],
-    [
-      -3.371344418512474, -2.7388218597689638, -3.9019726695746444,
-      -3.2088254890146994, -3.4965075614664802, -3.4965075614664802,
-      -2.5669716028423046, -3.431969040328909, -3.9019726695746444,
-      -3.6396084051071536, -4.258647613513377, -3.3141860046725258,
-      -3.26011878340225, -4.59511985013459, -2.9856819377004897,
-      -3.3141860046725258, -4.59511985013459, -4.007333185232471,
-      -2.3978952727983707, -3.0690635466395406, -4.258647613513377,
-      -4.258647613513377, -3.2088254890146994, -4.59511985013459,
-      -4.59511985013459, -4.59511985013459, -4.125116220888854,
-      -3.371344418512474, -4.258647613513377, -4.59511985013459,
-      -4.59511985013459, -4.59511985013459, -4.59511985013459,
-      -4.59511985013459, -4.59511985013459, -4.258647613513377,
-      -4.59511985013459, -4.258647613513377, -4.59511985013459,
-    ],
-  ],
-  'thresh': 0.029769981539181603,
-};
