+++ reporting/test/request/index.spec.js
@@ -316,7 +316,7 @@ describe('request/index', function () {
         parentFrameId: -1,
         method: 'GET',
         type: 'xmlhttprequest',
-        url: `http://tracker.com/track;uid=cliqz.com/tracking&uid2=cliqz.com/tracking&uid=${uid}?uid2=${uid}&encuid=${encodeURIComponent(
+        url: `http://tracker.com/track;uid=cliqz.com/tracking?uid2=cliqz.com/tracking&uid=${uid}?uid2=${uid}&encuid=${encodeURIComponent(
           uid,
         )}`,
         requestHeaders: mockRequestHeaders,
